+++ .github/workflows/test.yml
@@ -8,7 +8,8 @@ on:
   workflow_dispatch:
 
 jobs:
-  test:
+  unit-tests:
+    name: Unit Tests
     runs-on: ubuntu-latest
     strategy:
       fail-fast: false
@@ -30,15 +31,49 @@ jobs:
         pip install -e .
         pip install pytest pytest-cov
 
-    - name: Run tests with pytest
+    - name: Run unit tests with pytest
       run: |
-        pytest tests/ --doctest-modules --junitxml=junit/test-results-${{ matrix.python-version }}.xml
+        pytest tests/unit/ --doctest-modules --junitxml=junit/unit-test-results-${{ matrix.python-version }}.xml
     
-    - name: Upload pytest test results
-      # Use always() to always run this step to publish test results when there are test failures
+    - name: Upload unit test results
       if: ${{ always() }}
       uses: actions/upload-artifact@v4
       with:
-        name: pytest-results-${{ matrix.python-version }}
-        path: junit/test-results-${{ matrix.python-version }}.xml
+        name: unit-test-results-${{ matrix.python-version }}
+        path: junit/unit-test-results-${{ matrix.python-version }}.xml
+        retention-days: 7
+
+  integration-tests:
+    name: Integration Tests
+    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: ["3.10", "3.12", "3.13"]
+
+    steps:
+    - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4.2.2
+
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@8d9ed9ac5c53483de85588cdf95a591a75ab9f55 # v5.5.0
+      with:
+        python-version: ${{ matrix.python-version }}
+        cache: 'pip'
+
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install -e .
+        pip install pytest pytest-cov
+
+    - name: Run integration tests with pytest
+      run: |
+        pytest tests/integration/ --doctest-modules --junitxml=junit/integration-test-results-${{ matrix.python-version }}.xml
+    
+    - name: Upload integration test results
+      if: ${{ always() }}
+      uses: actions/upload-artifact@v4
+      with:
+        name: integration-test-results-${{ matrix.python-version }}
+        path: junit/integration-test-results-${{ matrix.python-version }}.xml
         retention-days: 7
